requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.10.5
aiosignal==1.3.2
attrs==25.1.0
frozenlist==1.5.0
idna==3.10
loguru==0.7.2
multidict==6.1.0
prometheus_client==0.21.0
propcache==0.2.1
PyYAML==6.0.2
websockets==15.0
yarl==1.18.3
